add nixpkgs 22.05 as a direct flake input
Instead of assuming users of this library will pass their own 22.05
version of nix, we simply have it specified in this flake itself. Rarely
will users want to specify their own instance of this, and this way, we
fix issues with inputs. This should also be a relatively stable input,
so I don't expect to run into many issues.